Feat/card domain model

+++ src/main/java/com/example/thirdtool/Deck/domain/model/Deck.java
@@ -1,6 +1,6 @@
 package com.example.thirdtool.Deck.domain.model;
 
-import com.example.thirdtool.Card.domain.model.Card;
+import com.example.thirdtool.LegacyCard.Card.domain.model.Card;
 import com.example.thirdtool.User.domain.model.UserEntity;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
@@ -9,6 +9,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -27,7 +28,7 @@ public class Deck {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private BigInteger id;
 
     @Column(nullable = false, length = 100)
     private String name;
